Cython implementation of the alignment algorithms
The Python and C implementations have been replaced by a C/Cython
version which is usable by Python as normal. The scoring function
included in the call to align is supposed to be a pure Python
function.

Currently a C library is required to be built first which is done
by running 'make install' inside the lib directory. After this is
done the Cython files are built by running
'python setup.py build_ext --inplace'.